u-boot-sdk: Use OpenBMC fork
This is a fork that is based on ASPEED's 2019.04 SDK fork. The goal is
to move to a mainline based tree, but while that work is underway, this
allows ast2600 support to move forward.

This gains support for NC-SI, and uses kconfig for the default
boot command allowing us to drop the patch to aspeed-common.h.

Machines will need to indicate which device tree to use. If possible we
should look to having a few common device trees, but for now they are
machine specific.

The Tacoma system has a single NCSI network device with eMMC and SPI NOR
storage devices. Rainier has two network devices with eMMC and SPI NOR.

Change-Id: I5b6925208550f687abe771523e07cdfa24c8620b
Signed-off-by: Joel Stanley <joel@jms.id.au>