ibm: Fix avsbus
Currently when building for p10bmc we see a warning about avsbus:

 WARNING: meta-ibm/recipes-phosphor/chassis/avsbus-control.bb: Unable
 to get checksum for avsbus-control SRC_URI entry avsbus-enable.sh:
 file could not be found

The p10bmc target doesn't have a script, becuase it doesn't use this
mechanism. It's specific to p9 machines, so modify the rule to only be
used for the ibm-power9-cpu machine.

The disable script is common between all implementations, so we can use
the one copy all machines that require it.

According to commit 863e5de9bbbd ("Swift: Put in avsbus-enable")
swift doesn't need the disable script. Therefore the systemd rule to run
it is not installed.

Change-Id: Id56349328b14972ae4f7c33a3b38b155e0ce3047
Signed-off-by: Joel Stanley <joel@jms.id.au>